requirements.txt
requests>=2.28.0
pytest>=7.0.0
psycopg2-binary>=2.9.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
